requirements.txt
# Web框架
flask==2.0.1
flask-restful==0.3.9
flask-cors==3.0.10

# 监控和指标
prometheus-client==0.11.0
influxdb-client==1.24.0
psutil==5.8.0

# 数据处理
pandas==1.3.3
numpy==1.21.2

# 安全
cryptography==3.4.7
pyOpenSSL==20.0.1
requests==2.26.0

# 日志
python-json-logger==2.0.2
loguru==0.5.3

# 配置
pyyaml==5.4.1
python-dotenv==0.19.0

# 测试
pytest==6.2.5
pytest-cov==2.12.1

# 工具
click==8.0.1
tqdm==4.62.3 